changed demo code structure and added many options
