--- src/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Copyright: 2010 RobotCub Consortium
-# Author: Sean Ryan Fanello, Carlo Ciliberto
-# CopyPolicy: Released under the terms of the GNU GPL v2.0.
-cmake_minimum_required(VERSION 2.6)
-
-#add_subdirectory(showOnTheFly)
-add_subdirectory(managerOnTheFly)
-
